Initial commit for nftables and serviceuser
Initial commit for the nftables firewall and serviceuser features.

nftables firewall is enabled by enabling the kernel configuration for
it, and creating a packagegroup that contains some core directives. This
allows creating a functional firewall.

Serviceuser feature allows creating a login user to the system. This
also disables the root login, and performs a check that the insecure
image features are not enabled.