linux-yocto: Add kernel hardening kmeta
This commit adds an in-source kmeta that can be used to enable kernel
hardening features. As the first feature, some attack surface reduction
is done. This is still not a complete list of the attack surface
reduction features.

The configuration items present in the cfg files are based on the raport
from kernel-hardening-checker tool.