distro: Add banner about restricted system
Add a banner that is shown before login to warn users that this is a
restricted system and unauthorized access is prohibited. Remove the
distro version from the net banner shown for users logging in through
SSH to avoid leaking information. For the users using serial the version
can still be shown.