useradd: Clean up the backup files
When creating or modifying users or groups with useradd, usermod,
groupadd, etc. a backup file with hyphen suffix is created. However, as
the final output of the build is going to be an initial firmware image,
having these backups makes no sense. The final image is supposed to be
the initial starting state, and there is no need for previous history.
Remove the backup files from the system to keep the initial state clean.