base-files: Harden /proc and tmpfs mounts
Add a new fstab for hardening the mounts of /proc and two tmpfs,
/var/volatile and /run.

For /proc, add hidepid=2 to hide process information from other users
than the current user.

For tmpfs mounts, ensure nosuid, nodev, and noexec options are present.

These may have unintended negative side effects for the legitimate
scripts and programs (especially the noexec option). In this kind of a
situation it should be considered if the program or script can be edited
to run without issues, and if not, the options should be adjusted.