shadow: Set password min and max age
Set the minimum and maximum ages for the password with the
SULKA_EXTRA_COMPLIANCY option. This is a thing that is commonly still
checked by audit tools, but in recent years there's been discussion
whether this is a good practice or not. It can improve the system
security, but more emphasis should be put on robust password rotation.