distro: Set nologin shell in the recipes
The nologin shell cannot be set in the main distro configuration,
because the packages installing users are not necessarily always
installed. Fix the system user login shell for the two commonly used
services, openssh and dbus, in their recipes by overwriting the useradd
params. The only thing that should be changed is the login shell. Add a
comment indicating this for future upgrade activities.