readme: Update quick start with kas workflow
Instead of having the user to manually jump through all kinds of hoops
to build the Sulka on top of Poky repository, use kas to build the image
with fewer manual steps. This should make it easier and less error prone
to get started with Sulka and to configure the distro.