pam, shadow: Use yescrypt for password encryption
Instead of using default SHA 512 for encrypting the passwords, use
yescrypt. Yescrypt is designed for password hashing and has a good
support available, making it a better choice for resisting password
brute-forcing attacks.

The cost factor of 8 was chosen as the default because it is
sufficiently high and it does not crash the test environment. The
authentication process is fairly slow, but this is considered an
acceptable tradeoff as the logins do not happen that often. If your
device cannot handle the cost factor, consider lowering it. If on the
other hand it seems too low, you can increase it. Remember to update
both common-password and login.defs.

Note that with newer versions of libpam it should be possible to read
the rounds for yescrypt from login.defs. When the libpam gets updated,
it should be checked if it is possible to read the pam_unix.so
configuration from login.defs to have centralized password encryption
configuration.