distro: Set serviceuser home to root directory
To avoid having serviceuser writable locations in the system, define the
root directory / as the home for the serviceuser. User still cannot
write to it, but HOME is defined when the user logs in.

This is almost the same as running useradd with -M flag (=do not create
the user's home directory), but with -M the HOME variable is not
defined. It seems that currently this gets automatically set to /, which
is the same as -d / option. However, this behavior may change in the
future, so it is better to explicitly set home to root.