lynis, scap-security-guide: Add recipes
Add updated recipes for lynis and scap-security-guide. These are usually
provided from the meta-security, but they are broken and the maintenance
of the meta-security is currently in transition phase.

The recipes are the same as in the meta-security, only SRCREV, PV and
checksums are updated, so they may or may not work perfectly.