nftables-configuration: Add example configurations
Add a few example configurations to the nftables-configuration recipe,
as the basic "drop everything" configuration has to be usually edited
anyway. Keep the default behavior still the same. Add a mechanism for
selecting one of the templates with a configuration variable. This
should help development work, and the templates should also assist in
writing the custom nftables configurations.