distro: Add AIDE
Add AIDE for intrusion and filesystem modification detection. Enable
AIDE by adding it to the Sulka packagegroup, and add an example
configuration for it.

Note that the AIDE database gets generated during the first boot.
However, after it has been generated, it still has to be moved in place,
i.e. the .new has to be removed from the database name.