sulka-rootfs-postprocess: Harden cron directories
Set the the cron directories to be readable only by root in a rootfs
postprocess task. This should not be done in a recipe, because multiple
packages may install files to the cron directories, and this would mean
that every recipe that installs cron tasks would have to be modified.
With a rootfs postprocess task this has to be done only once.