distro: Create home directory for service user
Create a home directory for the service user, and add it to the useradd
command. However, do not use the useradd command to create the home
directory, as it includes extra files from the template directory, such
as .profile and .bashrc. Create the home directory as a rootfs post
process task, and ensure that the user does not have write permissions
to the home directory.

Home directory is required as certain services (such SSH) attempt to
read files from it, but placing these files on the root of the file
system is not a good idea.