distro: Add configuration option to install SSH keys
Add SULKA_INSTALL_SSH_KEYS option that can be used to install SSH key
information to the firmware image. This is done with a rootfs
postprocess task that echoes the public key information from the given
directory into the authorized_keys file in the home directory of the
service user.

This commit allows easier installation of the SSH keys to the built
images, which in turn allows remote connections to the device.

Note that you should carefully consider installing the keys to the
firmware images. In case the private key leaks, it makes all the devices
running the same firmware image vulnerable. If possible, prefer to
generate the keys per-device.