audit: Run audit-rules service after postinsts
Sulka modifies the initial audit rules based on the filesystem content
in a postinstall task. This is done in a postinstall task, because it is
not possible for the audit recipe to know during build time what
binaries and directories are in the final image. Therefore, we need to
run the audit-rules.service after the post installation tasks have
finished. Otherwise the audit-rules.service fails on the first boot
before the rules are fixed.