distro: Use rsyslog as logging daemon
Instead of the default busybox syslogd logging daemon, use rsyslog.
rsyslog provides features that are not available in busybox syslog, like
signing logs and secure log transfer.

Another alternative that was considered was syslog-ng. The two were
considered to be mostly equal, but better performance and modularity of
the rsyslog lead to it getting selected.

However, if the user of the distro wishes to use some other logging
daemon, this assignment is only a weak default assignment.