refpolicy: Allow serial access for nft command
Allow iptables_t to read and write to the tty_device_t. Without this,
there will be deny messages in the audit logs if nft command is executed
from a context where ttyS0 is not unconfined (e.g. init script).
However, despite this deny, the command still seems to be successful.

Add new file configuration for systemd service nftables.service. This
has the same context as iptables.service.