linux: Remove linux kernel metadata
Remove the Linux kernel metadata from the distro meta-layer. This
metadata shall be moved to the meta-sulka-kernel layer. This move allows
separation of the responsibilities. meta-sulka-distro should focus
mostly on hardening the userspace, and enabling the features that
primarily affect the userspace (however, these may affect kernel
configuration, like SELinux). meta-sulka-kernel should focus on creating
a hardened kernel.

The development continues with the assumption that the meta-sulka-kernel
is used as the kernel provider, but this change should allow easier
swapping of the kernels if for example a board vendor kernel has to be
used.