distro: Skip uutils-coreutils recipe
Skip the uutils-coreutils recipe, as it prints out an error about
missing clang-native if SELinux is enabled and meta-clang is not in
bblayers.conf. This error message pops up on every build, even if the
uutils-coreutils is not being built, so it is disruptive. Add
instructions how the uutils-coreutils can be re-enabled if so required.