distro: Check cron directories before chmod
Check the existence of the cron directories before attempting to harden
them with chmod. Failure to check these directories results in the build
erroring out if the directory is not present. Cronie can be removed from
the build by user, and for example initramfs images do not install cron,
so this check is mandatory.