audit: Backport audit 4.1.2 from meta-openembedded
Backport the audit version 4.1.2 from meta-openembedded layer. The
version present in scarthgap, 4.0.1, does not seem to handle permission
filtering correctly. For example, if perm=wa is used in a rule, on
aarch64 architecture this incorrectly tries to audit rename syscall,
which does not exist. In newer version the filtering works correctly.

To backport the recipe, S had to be defined as ${WORKDIR}/git and
${UNPACKDIR} replaced with ${WORKDIR}. Other than that, the recipe is
left as-is.