distro: Recommend bash as the default shell
Bash is used as the default shell in testing, as oscap and
kernel-hardening-checker have it as a runtime dependency. Recommend
using it as the default shell in the distro as well to narrow the
difference between testing image and "actual" image. Using shell other
than Busybox shells also fixes the problem with SELinux Busybox wrappers
that cause the arguments for the login shell to be lost and login shell
not getting correctlydetected. See the following issue for more
information:

https://codeberg.org/AltidSec/meta-sulka-distro/issues/42

However, this is not really mandatory for the operation of the distro,
and the Bash inflates the image size a bit, so only recommend installing
it.