audit: Generate the rules in postinst
Generate the rules on the device in the postinst task after the rules
have been finalized on the device. This allows having a proper ruleset
soon after starting the device, instead of relying on the empty default
set provided by the recipe.

Note that this behavior is problematic from the read-only file system
point of view.