distro: Use packagegroup-selinux-minimal with SELinux
Use the packagegroup-selinux-minimal to provide SELinux tools instead of
packagegroup-core-selinux. Minimal toolset contains fewer tools, but
should still be enough for most use-cases. Core toolset can be used in
development images.

Remove coreutils from the minimal toolset, as it is GPLv3 licensed.
uutils-coreutils shall be used as the preferred provider for the utils.
In addition, remove broken fixfiles script that relied on bashism and
bash is GPLv3 licensed. This in practice prevents using
FIRST_BOOT_RELABEL, so print a fatal error if it is enabled.