distro: Change default syslogger to syslog-ng
As rsyslog is GPLv3 licensed, it should be removed from the image.
Syslog-ng has a more permissive license, and it supports the features
required from a secure logging daemon (secure remote logging, log
sealing).

By default, there is a syslog.conf conversion plugin the requires awk,
and a debug bundle generator that uses awk. Remove these to get rid of
the gawk dependency, as that is GPLv3 licensed as well.