packagegroup-sulka-monitoring: Remove acct & aide
Remove acct and aide packages from the monitoring packagegroup. They are
GPLv3 licensed, making their usage in embedded contexts somewhat
difficult. Also, these two tools required the most manual setup, because
aide does not perform checks automatically and acct generates a lot of
logs, so their removal should result in a simpler system. Keep the acct
recipe in case it is needed or there is some use-case for it.