audit: Check files before editing them in postinstall
Check the files that the postinstall task modifies and executes before
attempting to access them. Audit can be installed separately from the
auditd, and in that situation the rule files are not present.

It would be more correct to add this postinstall task to the auditd
package instead of audit package. However, auditd installs systemd
services that cannot be restarted, but the default postinstall task
attempts to restart them, resulting in a failure if performed during
first boot. Therefore, this postinstall task is in the main audit
package and checks are performed to ensure that no actions are taken if
auditd is not actually installed.