refpolicy-targeted: Reference policy fixes
Fix the reference policy to allow most of the actions requested by
syslog-ng, and replace the rsyslog patch with the syslog-ng patch. The
only non-allowed tasks are network access (as was with rsyslog) and root
write, which seems to be quite dangerous.

In addition, fix dmesg proc access related issues. Dmesg should be
trustworthy enough to allow this access.