syslog-ng: Disable xconsole logging
xconsole logging isn't that useful in embedded environments and would
required loosening the SELinux policy, so disable it altogether. Note
that when moving from sysvinit to systemd this patch should be dropped,
as it patches sysvinit services which won't be required anymore at that
stage. The configuration file could be supplied by the distro.